make ./gradlew verify the gradle binary it downloads
This just makes sure that the downloaded binary matches the official
release binary, i.e. has no malware in it.  Also, the -all package is not
needed to run the build, only the smaller -bin is required.

The gradle-wrapper.jar was copied from the open-keychain project, since
they know this stuff better than me :)
https://github.com/open-keychain/open-keychain/raw/92d2a052f1e660f498294bdb57f1237ae81fc770/gradle/wrapper/gradle-wrapper.jar