remove Android Studio files from git
everything in .idea/ and all .iml files should not be included in git. They
are generated by Android Studio when opening a gradle project.  They will
be different for each person's Android Studio, so if they are in git, it'll
be an endless source of git conflicts.

Android Studio will recreate all of these files from the gradle files if
you do "Open an Existing Android Studio project"